FIX:  - sending big messages via WS  - system crashes Added:  - autotasks - functionality which allow user to start and kill scripts with FC  - all important thread calls are using Workers Project structure cleaned

+++ core/hardware/machine_info.c
@@ -314,13 +314,13 @@ char* GetSystemUniqueId()
 		strcat( uniqueID, mname );
 	}
 
-	DEBUG("uniqueID %s len %d\n", uniqueID, strlen(uniqueID) );
+	DEBUG("uniqueID %s len %d\n", uniqueID, (int)strlen(uniqueID) );
 	
 #ifndef CYGWIN_BUILD
 	FUWORD *id = ComputeSystemUniqueId();
 	if( id != NULL )
 	{
-		DEBUG("id is not null uniqueID %s len %d\n", uniqueID, strlen(uniqueID) );
+		DEBUG("id is not null uniqueID %s len %d\n", uniqueID, (int) strlen(uniqueID) );
 		unsigned int i = 0;
 		
 		for ( i = 0; i < 5; i++ )
